+++ client/client.go
@@ -363,6 +363,26 @@ func (c *Client) Flush() error {
 	return c.ok(c.rdr)
 }
 
+// List queues explicitly or use "*" to pause all known queues
+func (c *Client) PauseQueues(names ...string) error {
+	err := c.writeLine(c.wtr, "QUEUE PAUSE", strings.Join(names, " "))
+	if err != nil {
+		return err
+	}
+
+	return c.ok(c.rdr)
+}
+
+// List queues explicitly or use "*" to resume all known queues
+func (c *Client) ResumeQueues(names ...string) error {
+	err := c.writeLine(c.wtr, "QUEUE RESUME", strings.Join(names, " "))
+	if err != nil {
+		return err
+	}
+
+	return c.ok(c.rdr)
+}
+
 func (c *Client) Info() (map[string]interface{}, error) {
 	err := c.writeLine(c.wtr, "INFO", nil)
 	if err != nil {