+++ Changes.md
@@ -4,7 +4,7 @@ Changelog: Faktory || [Faktory Pro](https://github.com/contribsys/faktory/blob/m
 
 ## HEAD
 
-- Add queue pausing / unpausing to /queues page [#336]
+- Add `QUEUE PAUSE` and `QUEUE RESUME` commands, add buttons to the /queues page [#336]
 - The Busy page can now display current RSS for Worker processes [#339]
 - The /debug page now displays Redis latency [#337]
 - Add `Dialer` support for custom `faktory.Client` connections [#330]